fix(LinkBarSearchField): switch to dorris icons
