+++ src/atoms/SvgIcon/dorris/ArrowDown.js
@@ -4,7 +4,7 @@ import Svg from './Svg';
 
 const ArrowDown = props => (
 	<Svg {...props}>
-		<polyline points="31.8,11.8 24.1,19.4 16.5,27 8.8,19.4 1.2,11.8 " />
+		<polyline points="31.8,11.8 24.1,19.4 16.5,27 8.8,19.4 1.2,11.8" />
 	</Svg>
 );
