+++ src/utils/index.js
@@ -4,3 +4,4 @@ export { getVariable } from './get-variable';
 export { unshadeColorString } from './unshade-color-string';
 export { isBright, isDark } from './luminance';
 export { calculateTextColorFromName, calculateTextColorFromTheme } from './calculate-text-color';
+export { linkBarElementSizes } from './link-bar-element-sizes';
